refactor: split subtitle OCR helpers and tests into modules
