+++ src/subtitle_ocr.rs
@@ -31,6 +31,14 @@ use std::sync::{
 };
 
 use crate::{OcrEngine, OcrFormat, SubMode};
+use text_processing::{
+    is_english_language, language_uses_spaces, lines_text_for_quality, ocr_text_quality_score,
+    postprocess_ocr_text, ppocr_average_confidence, ppocr_needs_quality_fallback,
+};
+#[cfg(test)]
+use text_processing::{ppocr_spacing_needs_fallback, split_glued_ascii_token};
+
+mod text_processing;
 
 #[derive(Debug, Clone)]
 pub struct OcrSubtitleTrack {
@@ -2047,1094 +2055,6 @@ fn extract_subtitle_lines(
     Ok((lines, had_imagery))
 }
 
-fn language_uses_spaces(language: &str) -> bool {
-    let lang = language.to_lowercase();
-    matches!(
-        lang.as_str(),
-        "eng"
-            | "en"
-            | "en-us"
-            | "en_us"
-            | "fre"
-            | "fra"
-            | "fr"
-            | "spa"
-            | "es"
-            | "de"
-            | "deu"
-            | "ger"
-            | "it"
-            | "ita"
-            | "pt"
-            | "por"
-            | "nl"
-            | "nld"
-            | "sv"
-            | "swe"
-            | "da"
-            | "dan"
-            | "no"
-            | "nor"
-            | "fi"
-            | "fin"
-    )
-}
-
-fn ppocr_spacing_needs_fallback(lines: &[OcrLine]) -> bool {
-    if lines.is_empty() {
-        return false;
-    }
-    let mut has_spaces = false;
-    let mut long_token = false;
-    let mut has_letters = false;
-    for line in lines {
-        let text = line.text.trim();
-        if text.contains(' ') {
-            has_spaces = true;
-            break;
-        }
-        if text.len() >= 12 {
-            long_token = true;
-        }
-        if text.chars().any(|c| c.is_alphabetic()) {
-            has_letters = true;
-        }
-    }
-    has_letters && long_token && !has_spaces
-}
-
-fn postprocess_ocr_text(text: &str, language: &str) -> String {
-    let mut out = normalize_utf8_text(text);
-    if out.is_empty() {
-        return out;
-    }
-
-    if !is_english_language(language) {
-        return out;
-    }
-
-    out = normalize_english_ocr_confusions(&out);
-    out = insert_space_after_punctuation(&out);
-    out = insert_space_between_letters_and_digits(&out);
-    out = insert_space_before_opening_quote(&out);
-    out = split_glued_english_phrases(&out);
-
-    // Targeted corrections for frequently observed OCR glue patterns.
-    const ENGLISH_GLUE_FIXES: [(&str, &str); 42] = [
-        ("noneother", "none other"),
-        ("notonlyme", "not only me"),
-        ("notonly", "not only"),
-        ("itis", "it is"),
-        ("whylost", "why lost"),
-        ("hesalive", "he's alive"),
-        ("he'salive", "he's alive"),
-        ("thats", "that's"),
-        ("goodwork", "good work"),
-        ("burnit", "burn it"),
-        ("yessir", "yes sir"),
-        ("praisetoyou", "praise to you"),
-        ("lordjesuschrist", "Lord Jesus Christ"),
-        ("paxchristi", "pax Christi"),
-        ("praisedbegod", "praised be God"),
-        ("constablecrane", "constable crane"),
-        ("whathappenedtohim", "what happened to him"),
-        ("beforehewentintotheriver", "before he went into the river"),
-        (
-            "thereislittlepeaceinthislandnow",
-            "there is little peace in this land now",
-        ),
-        ("allourprogresshasended", "all our progress has ended"),
-        ("newsuffering", "new suffering"),
-        ("tobeasdarkasitisnow", "to be as dark as it is now"),
-        ("tobeasdarkasit isnow", "to be as dark as it is now"),
-        ("to be as dark as itis now", "to be as dark as it is now"),
-        ("an done of", "and one of"),
-        ("butit's", "but it's"),
-        ("isit?", "is it?"),
-        (
-            "andthepainwouldbeprolonged",
-            "and the pain would be prolonged",
-        ),
-        (
-            "eachsmallsplashofthewater",
-            "each small splash of the water",
-        ),
-        ("waslikeaburningcoal", "was like a burning coal"),
-        ("therearehotspringsthere", "there are hot springs there"),
-        ("toabandongod", "to abandon God"),
-        (
-            "sotheycoulddemonstratethestrengthoftheirfaith",
-            "so they could demonstrate the strength of their faith",
-        ),
-        (
-            "andthepresenceofgodwithinthem",
-            "and the presence of God within them",
-        ),
-        ("standdown", "stand down"),
-        ("loppedoff", "lopped off"),
-        ("ibegpardon", "I beg pardon"),
-        ("ihavenot", "I have not"),
-        ("ishall", "I shall"),
-        ("begpardon", "beg pardon"),
-        ("havenot", "have not"),
-        ("l9th", "19th"),
-    ];
-    for (from, to) in ENGLISH_GLUE_FIXES {
-        out = replace_case_insensitive_ascii(&out, from, to);
-    }
-
-    normalize_utf8_text(&out)
-}
-
-fn is_english_language(language: &str) -> bool {
-    let lang = language.trim().to_ascii_lowercase();
-    matches!(lang.as_str(), "eng" | "en" | "en-us" | "en_us")
-}
-
-fn normalize_english_ocr_confusions(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let mut token = String::new();
-    let flush_token = |tok: &mut String, out: &mut String| {
-        if tok.is_empty() {
-            return;
-        }
-        let has_alpha = tok.chars().any(|c| c.is_ascii_alphabetic());
-        let has_digit = tok.chars().any(|c| c.is_ascii_digit());
-        let mut normalized = tok.clone();
-        if has_alpha && has_digit {
-            normalized = normalized
-                .replace('0', "o")
-                .replace('1', "l")
-                .replace('5', "s")
-                .replace('8', "b");
-        }
-        let normalized_lc = normalized.to_ascii_lowercase();
-        if let Some(rest) = normalized_lc.strip_prefix('l') {
-            if rest.chars().next().is_some_and(|ch| ch.is_ascii_digit())
-                && (rest.ends_with("st")
-                    || rest.ends_with("nd")
-                    || rest.ends_with("rd")
-                    || rest.ends_with("th"))
-            {
-                normalized.replace_range(0..1, "1");
-            }
-        }
-        normalized = normalized.replace('|', "I").replace("vv", "w");
-        out.push_str(&normalized);
-        tok.clear();
-    };
-
-    for ch in input.chars() {
-        if ch.is_ascii_alphanumeric() || ch == '\'' || ch == '|' {
-            token.push(ch);
-        } else {
-            flush_token(&mut token, &mut out);
-            out.push(ch);
-        }
-    }
-    flush_token(&mut token, &mut out);
-    out
-}
-
-fn split_glued_english_phrases(input: &str) -> String {
-    let mut out = String::with_capacity(input.len() + 8);
-    let mut token = String::new();
-
-    let flush_token = |tok: &mut String, out: &mut String| {
-        if tok.is_empty() {
-            return;
-        }
-        if let Some(split) = split_glued_ascii_token(tok) {
-            out.push_str(&split);
-        } else {
-            out.push_str(tok);
-        }
-        tok.clear();
-    };
-
-    for ch in input.chars() {
-        if ch.is_ascii_alphabetic() || ch == '\'' {
-            token.push(ch);
-        } else {
-            flush_token(&mut token, &mut out);
-            out.push(ch);
-        }
-    }
-    flush_token(&mut token, &mut out);
-
-    out
-}
-
-fn split_glued_ascii_token(token: &str) -> Option<String> {
-    if token.len() < 5 || !token.is_ascii() {
-        return None;
-    }
-    if !token
-        .chars()
-        .all(|ch| ch.is_ascii_alphabetic() || ch == '\'')
-    {
-        return None;
-    }
-
-    let lower = token.to_ascii_lowercase();
-    if let Some(split) = split_glued_contraction(token, &lower) {
-        return Some(split);
-    }
-    if is_common_english_word(&lower) {
-        return None;
-    }
-
-    if matches!(token.chars().next(), Some('I' | 'i')) && token.len() >= 5 {
-        const I_PREFIX_CONTINUATIONS: [&str; 16] = [
-            "am", "have", "had", "shall", "will", "beg", "think", "know", "need", "must", "want",
-            "did", "do", "was", "were", "would",
-        ];
-        let rest = &lower[1..];
-        if I_PREFIX_CONTINUATIONS
-            .iter()
-            .any(|prefix| rest.starts_with(prefix))
-        {
-            let split_rest =
-                segment_glued_english_token(&token[1..]).unwrap_or_else(|| token[1..].to_string());
-            return Some(format!("{} {}", &token[..1], split_rest));
-        }
-    }
-
-    for suffix in [
-        "down", "off", "out", "up", "in", "on", "over", "under", "away", "back",
-    ] {
-        if lower.ends_with(suffix) {
-            let split = token.len() - suffix.len();
-            if split >= 4 && is_common_english_word(&lower[..split]) {
-                return Some(format!("{} {}", &token[..split], &token[split..]));
-            }
-        }
-    }
-
-    segment_glued_english_token(token)
-}
-
-fn segment_glued_english_token(token: &str) -> Option<String> {
-    let lower = token.to_ascii_lowercase();
-    if lower.len() < 5 || is_common_english_word(&lower) {
-        return None;
-    }
-
-    // Dynamic programming split over common English words.
-    let n = lower.len();
-    let mut best: Vec<Option<(i32, usize, usize)>> = vec![None; n + 1]; // (score, prev_idx, segments)
-    best[0] = Some((0, 0, 0));
-    for end in 1..=n {
-        let start_min = end.saturating_sub(12);
-        for start in start_min..end {
-            let Some((prev_score, _prev_idx, prev_segments)) = best[start] else {
-                continue;
-            };
-            let candidate = &lower[start..end];
-            if !is_common_english_word(candidate) {
-                continue;
-            }
-            let segment_len = end - start;
-            let score = prev_score + (segment_len as i32 * segment_len as i32) - 4;
-            let segments = prev_segments + 1;
-            let should_replace = best[end]
-                .as_ref()
-                .map(|(current_score, _, current_segments)| {
-                    score > *current_score
-                        || (score == *current_score && segments < *current_segments)
-                })
-                .unwrap_or(true);
-            if should_replace {
-                best[end] = Some((score, start, segments));
-            }
-        }
-    }
-
-    let (_score, _prev, segment_count) = best[n]?;
-    if segment_count < 2 {
-        return None;
-    }
-
-    let mut pieces = Vec::new();
-    let mut idx = n;
-    while idx > 0 {
-        let (_score, prev_idx, _segments) = best[idx]?;
-        pieces.push((prev_idx, idx));
-        idx = prev_idx;
-    }
-    pieces.reverse();
-
-    // Guard against pathological over-segmentation (e.g. many tiny tokens).
-    let avg_segment_len = n as f32 / pieces.len() as f32;
-    if pieces.len() >= 5 && avg_segment_len < 2.6 {
-        return None;
-    }
-    if pieces.iter().any(|(start, end)| {
-        end - start == 1 && &lower[*start..*end] != "i" && &lower[*start..*end] != "a"
-    }) {
-        return None;
-    }
-
-    Some(
-        pieces
-            .into_iter()
-            .map(|(start, end)| token[start..end].to_string())
-            .collect::<Vec<_>>()
-            .join(" "),
-    )
-}
-
-fn split_glued_contraction(token: &str, lower: &str) -> Option<String> {
-    let apostrophe = token.find('\'')?;
-    if apostrophe == 0 || apostrophe + 1 >= token.len() {
-        return None;
-    }
-    if token[apostrophe + 1..].contains('\'') {
-        return None;
-    }
-
-    const SUFFIXES: [&str; 7] = ["s", "re", "ve", "ll", "d", "m", "t"];
-    for suffix in SUFFIXES {
-        let suffix_start = apostrophe + 1;
-        if !lower[suffix_start..].starts_with(suffix) {
-            continue;
-        }
-        let contraction_len = suffix_start + suffix.len();
-        if contraction_len + 2 > token.len() {
-            return None;
-        }
-        let rest = &token[contraction_len..];
-        if !rest.chars().all(|ch| ch.is_ascii_alphabetic()) {
-            return None;
-        }
-        let split_rest = segment_glued_english_token(rest).unwrap_or_else(|| rest.to_string());
-        return Some(format!("{} {}", &token[..contraction_len], split_rest));
-    }
-    None
-}
-
-fn is_common_english_word(word: &str) -> bool {
-    const WORDS: [&str; 511] = [
-        "a",
-        "abandon",
-        "able",
-        "about",
-        "above",
-        "according",
-        "across",
-        "actually",
-        "add",
-        "after",
-        "again",
-        "against",
-        "ago",
-        "ahead",
-        "air",
-        "alive",
-        "all",
-        "allow",
-        "almost",
-        "along",
-        "already",
-        "also",
-        "although",
-        "always",
-        "am",
-        "among",
-        "amount",
-        "an",
-        "and",
-        "another",
-        "any",
-        "anyone",
-        "anything",
-        "apostatize",
-        "apostatized",
-        "are",
-        "around",
-        "as",
-        "ask",
-        "asked",
-        "at",
-        "away",
-        "back",
-        "bad",
-        "base",
-        "be",
-        "became",
-        "because",
-        "become",
-        "becoming",
-        "been",
-        "before",
-        "beg",
-        "behind",
-        "being",
-        "below",
-        "beside",
-        "best",
-        "better",
-        "between",
-        "black",
-        "blood",
-        "blue",
-        "body",
-        "book",
-        "born",
-        "both",
-        "break",
-        "bring",
-        "brother",
-        "brought",
-        "build",
-        "built",
-        "burn",
-        "burning",
-        "but",
-        "by",
-        "call",
-        "came",
-        "can",
-        "cannot",
-        "cause",
-        "century",
-        "children",
-        "christ",
-        "christi",
-        "christians",
-        "church",
-        "city",
-        "clear",
-        "close",
-        "coal",
-        "cold",
-        "come",
-        "constable",
-        "continue",
-        "could",
-        "country",
-        "courage",
-        "crane",
-        "cut",
-        "dandelion",
-        "dark",
-        "day",
-        "days",
-        "dead",
-        "death",
-        "deep",
-        "demonstrate",
-        "denounced",
-        "despite",
-        "did",
-        "die",
-        "different",
-        "do",
-        "does",
-        "done",
-        "door",
-        "down",
-        "drops",
-        "during",
-        "dutch",
-        "each",
-        "early",
-        "earth",
-        "easy",
-        "either",
-        "ended",
-        "enough",
-        "even",
-        "ever",
-        "every",
-        "face",
-        "fact",
-        "faith",
-        "family",
-        "far",
-        "father",
-        "fear",
-        "ferreira",
-        "few",
-        "figure",
-        "filled",
-        "finally",
-        "fire",
-        "first",
-        "five",
-        "floor",
-        "follow",
-        "food",
-        "for",
-        "force",
-        "form",
-        "four",
-        "free",
-        "friars",
-        "friend",
-        "from",
-        "front",
-        "full",
-        "game",
-        "get",
-        "girl",
-        "give",
-        "given",
-        "gives",
-        "glass",
-        "go",
-        "god",
-        "good",
-        "gospel",
-        "governor",
-        "great",
-        "green",
-        "group",
-        "grow",
-        "had",
-        "half",
-        "hand",
-        "happened",
-        "hard",
-        "has",
-        "have",
-        "he",
-        "head",
-        "hear",
-        "heart",
-        "heavy",
-        "hells",
-        "her",
-        "here",
-        "hes",
-        "hidden",
-        "him",
-        "his",
-        "history",
-        "holes",
-        "home",
-        "hope",
-        "hot",
-        "house",
-        "how",
-        "however",
-        "i",
-        "idea",
-        "if",
-        "important",
-        "in",
-        "inside",
-        "into",
-        "is",
-        "it",
-        "its",
-        "japan",
-        "japanese",
-        "jesus",
-        "job",
-        "just",
-        "keep",
-        "kind",
-        "king",
-        "knew",
-        "know",
-        "known",
-        "ladles",
-        "land",
-        "large",
-        "last",
-        "late",
-        "later",
-        "leave",
-        "left",
-        "let",
-        "letter",
-        "life",
-        "light",
-        "like",
-        "line",
-        "list",
-        "little",
-        "live",
-        "living",
-        "long",
-        "looked",
-        "looking",
-        "looks",
-        "lopped",
-        "lord",
-        "lost",
-        "love",
-        "low",
-        "made",
-        "main",
-        "make",
-        "man",
-        "many",
-        "may",
-        "me",
-        "mean",
-        "means",
-        "men",
-        "might",
-        "mind",
-        "minute",
-        "mockery",
-        "money",
-        "month",
-        "months",
-        "more",
-        "morning",
-        "most",
-        "mother",
-        "mountain",
-        "move",
-        "moved",
-        "must",
-        "my",
-        "nagasaki",
-        "name",
-        "near",
-        "need",
-        "never",
-        "new",
-        "news",
-        "next",
-        "night",
-        "no",
-        "none",
-        "nor",
-        "north",
-        "not",
-        "nothing",
-        "notice",
-        "now",
-        "number",
-        "of",
-        "off",
-        "officials",
-        "often",
-        "old",
-        "on",
-        "once",
-        "one",
-        "only",
-        "open",
-        "or",
-        "order",
-        "other",
-        "others",
-        "our",
-        "out",
-        "outside",
-        "over",
-        "own",
-        "padres",
-        "pain",
-        "paper",
-        "pardon",
-        "part",
-        "partly",
-        "pass",
-        "past",
-        "pax",
-        "pay",
-        "peace",
-        "people",
-        "persecution",
-        "person",
-        "piece",
-        "place",
-        "point",
-        "portugal",
-        "possible",
-        "power",
-        "praise",
-        "praised",
-        "presence",
-        "priests",
-        "probably",
-        "progress",
-        "prolonged",
-        "proven",
-        "public",
-        "put",
-        "question",
-        "rain",
-        "ransomed",
-        "rather",
-        "reach",
-        "read",
-        "ready",
-        "really",
-        "red",
-        "refused",
-        "remain",
-        "remained",
-        "remember",
-        "repression",
-        "right",
-        "risked",
-        "river",
-        "road",
-        "room",
-        "run",
-        "said",
-        "same",
-        "saw",
-        "say",
-        "school",
-        "second",
-        "secret",
-        "see",
-        "seem",
-        "seen",
-        "service",
-        "set",
-        "several",
-        "shall",
-        "she",
-        "show",
-        "side",
-        "since",
-        "sir",
-        "six",
-        "slowly",
-        "small",
-        "smuggled",
-        "so",
-        "society",
-        "some",
-        "something",
-        "son",
-        "soon",
-        "sound",
-        "south",
-        "splash",
-        "spread",
-        "springs",
-        "stand",
-        "state",
-        "still",
-        "stop",
-        "stopped",
-        "story",
-        "strength",
-        "stronger",
-        "strongest",
-        "such",
-        "suffering",
-        "sure",
-        "surrendered",
-        "sweeping",
-        "table",
-        "taken",
-        "taking",
-        "talk",
-        "teacher",
-        "team",
-        "tell",
-        "ten",
-        "terms",
-        "text",
-        "than",
-        "that",
-        "thats",
-        "the",
-        "their",
-        "them",
-        "then",
-        "there",
-        "these",
-        "they",
-        "thing",
-        "things",
-        "think",
-        "this",
-        "those",
-        "though",
-        "thousands",
-        "three",
-        "through",
-        "time",
-        "to",
-        "today",
-        "together",
-        "told",
-        "tonight",
-        "too",
-        "took",
-        "top",
-        "tortured",
-        "toward",
-        "trader",
-        "traveling",
-        "true",
-        "truth",
-        "turn",
-        "two",
-        "under",
-        "until",
-        "unzen",
-        "up",
-        "upon",
-        "us",
-        "use",
-        "used",
-        "using",
-        "usually",
-        "value",
-        "very",
-        "view",
-        "voice",
-        "wait",
-        "wall",
-        "want",
-        "war",
-        "was",
-        "watch",
-        "water",
-        "we",
-        "week",
-        "well",
-        "went",
-        "were",
-        "what",
-        "when",
-        "where",
-        "which",
-        "white",
-        "who",
-        "whole",
-        "whose",
-        "why",
-        "will",
-        "with",
-        "within",
-        "woman",
-        "word",
-        "words",
-        "work",
-        "world",
-        "worse",
-        "would",
-        "wrote",
-        "year",
-        "years",
-        "yes",
-        "yet",
-        "you",
-        "young",
-        "your",
-    ];
-    WORDS.binary_search(&word).is_ok()
-}
-
-fn insert_space_after_punctuation(input: &str) -> String {
-    let mut out = String::with_capacity(input.len() + 8);
-    let chars: Vec<char> = input.chars().collect();
-    for (i, ch) in chars.iter().enumerate() {
-        out.push(*ch);
-        if matches!(ch, ',' | '.' | ';' | ':' | '!' | '?')
-            && chars
-                .get(i + 1)
-                .is_some_and(|next| next.is_ascii_alphabetic())
-        {
-            out.push(' ');
-        }
-    }
-    out
-}
-
-fn insert_space_between_letters_and_digits(input: &str) -> String {
-    let mut out = String::with_capacity(input.len() + 8);
-    let chars: Vec<char> = input.chars().collect();
-    for (i, ch) in chars.iter().enumerate() {
-        out.push(*ch);
-        if let Some(next) = chars.get(i + 1) {
-            let alpha_to_digit = ch.is_ascii_alphabetic() && next.is_ascii_digit();
-            let digit_to_alpha = ch.is_ascii_digit() && next.is_ascii_alphabetic();
-            let ordinal_suffix = if digit_to_alpha && i + 2 < chars.len() {
-                let a = chars[i + 1].to_ascii_lowercase();
-                let b = chars[i + 2].to_ascii_lowercase();
-                matches!((a, b), ('s', 't') | ('n', 'd') | ('r', 'd') | ('t', 'h'))
-                    && chars
-                        .get(i + 3)
-                        .map(|c| !c.is_ascii_alphabetic())
-                        .unwrap_or(true)
-            } else {
-                false
-            };
-            let boundary = alpha_to_digit || (digit_to_alpha && !ordinal_suffix);
-            if boundary && *ch != ' ' && *next != ' ' {
-                out.push(' ');
-            }
-        }
-    }
-    out
-}
-
-fn insert_space_before_opening_quote(input: &str) -> String {
-    let mut out = String::with_capacity(input.len() + 4);
-    let chars: Vec<char> = input.chars().collect();
-    for (i, ch) in chars.iter().enumerate() {
-        if *ch == '"'
-            && i > 0
-            && i + 1 < chars.len()
-            && chars[i - 1].is_ascii_alphabetic()
-            && chars[i + 1].is_ascii_alphabetic()
-            && !out.ends_with(' ')
-        {
-            out.push(' ');
-        }
-        out.push(*ch);
-    }
-    out
-}
-
-fn replace_case_insensitive_ascii(input: &str, from: &str, to: &str) -> String {
-    if from.is_empty() {
-        return input.to_string();
-    }
-    let input_lc = input.to_ascii_lowercase();
-    let from_lc = from.to_ascii_lowercase();
-    let mut out = String::with_capacity(input.len());
-    let mut pos = 0usize;
-    while let Some(rel_idx) = input_lc[pos..].find(&from_lc) {
-        let idx = pos + rel_idx;
-        out.push_str(&input[pos..idx]);
-        let orig = &input[idx..idx + from.len()];
-        let replacement = if orig
-            .chars()
-            .next()
-            .is_some_and(|ch| ch.is_ascii_uppercase())
-        {
-            let mut chars = to.chars();
-            if let Some(first) = chars.next() {
-                format!(
-                    "{}{}",
-                    first.to_ascii_uppercase(),
-                    chars.collect::<String>()
-                )
-            } else {
-                to.to_string()
-            }
-        } else {
-            to.to_string()
-        };
-        out.push_str(&replacement);
-        pos = idx + from.len();
-    }
-    out.push_str(&input[pos..]);
-    out
-}
-
-fn lines_text_for_quality(lines: &[OcrLine]) -> String {
-    normalize_utf8_text(
-        &lines
-            .iter()
-            .map(|line| line.text.trim())
-            .filter(|text| !text.is_empty())
-            .collect::<Vec<_>>()
-            .join(" "),
-    )
-}
-
-fn ppocr_average_confidence(lines: &[OcrLine]) -> Option<f32> {
-    let mut sum = 0.0f32;
-    let mut count = 0usize;
-    for score in lines.iter().filter_map(|line| line.score) {
-        if score.is_finite() {
-            sum += score;
-            count += 1;
-        }
-    }
-    if count == 0 {
-        None
-    } else {
-        Some(sum / count as f32)
-    }
-}
-
-fn ocr_text_quality_score(text: &str, language: &str) -> f32 {
-    let text = normalize_utf8_text(text);
-    if text.is_empty() {
-        return 0.0;
-    }
-
-    let mut letters = 0usize;
-    let mut digits = 0usize;
-    let mut spaces = 0usize;
-    let mut punctuation = 0usize;
-    let mut noise = 0usize;
-
-    for ch in text.chars() {
-        if ch.is_alphabetic() {
-            letters += 1;
-        } else if ch.is_ascii_digit() {
-            digits += 1;
-        } else if ch.is_whitespace() {
-            spaces += 1;
-        } else if ch.is_ascii_punctuation() || "“”‘’…".contains(ch) {
-            punctuation += 1;
-        } else {
-            noise += 1;
-        }
-    }
-
-    let total = (letters + digits + spaces + punctuation + noise).max(1) as f32;
-    let mut score = 1.0f32;
-
-    let noise_ratio = noise as f32 / total;
-    if noise_ratio > 0.0 {
-        score -= noise_ratio * 1.2;
-    }
-
-    if text.contains("@&") {
-        score -= 0.18;
-    }
-    if text.contains('|') {
-        score -= 0.12;
-    }
-
-    let words_vec = text.split_whitespace().collect::<Vec<_>>();
-    let word_count = words_vec.len().max(1);
-    let avg_word_len = letters as f32 / word_count as f32;
-    let long_word_count = words_vec.iter().filter(|word| word.len() >= 14).count();
-
-    if language_uses_spaces(language) {
-        if letters >= 12 && word_count <= 1 {
-            score -= 0.2;
-        }
-        if avg_word_len > 8.5 {
-            score -= 0.12;
-        }
-        if long_word_count > 0 {
-            score -= (long_word_count as f32 * 0.04).min(0.2);
-        }
-    }
-
-    if letters == 0 && digits == 0 {
-        score -= 0.3;
-    }
-
-    // Slightly reward candidates with enough character coverage.
-    let coverage_bonus = (letters as f32 / 24.0).min(0.2);
-    (score + coverage_bonus).clamp(0.0, 1.0)
-}
-
-fn ppocr_needs_quality_fallback(lines: &[OcrLine], language: &str) -> bool {
-    if lines.is_empty() {
-        return false;
-    }
-    if ppocr_spacing_needs_fallback(lines) {
-        return true;
-    }
-
-    let quality = ocr_text_quality_score(&lines_text_for_quality(lines), language);
-    quality < 0.45
-}
-
 #[derive(Debug, Clone)]
 struct TesseractCandidate {
     text: String,
@@ -4125,904 +3045,4 @@ fn is_image_based_subtitle(codec_id: ffi::AVCodecID) -> bool {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use mockito::Server;
-    use std::fs::File;
-    use std::io::Write;
-    use std::ptr;
-    use strsim::jaro_winkler;
-    use tempfile::TempDir;
-
-    fn normalize_text_for_word_similarity(input: &str) -> String {
-        input
-            .to_uppercase()
-            .split_whitespace()
-            .collect::<Vec<_>>()
-            .join(" ")
-    }
-
-    fn normalize_text_for_char_similarity(input: &str) -> String {
-        input
-            .to_uppercase()
-            .chars()
-            .filter(|c| !c.is_whitespace())
-            .collect()
-    }
-
-    #[allow(clippy::needless_range_loop)]
-    fn char_error_rate(expected: &str, actual: &str) -> f32 {
-        let expected_chars: Vec<char> = expected.chars().collect();
-        let actual_chars: Vec<char> = actual.chars().collect();
-        if expected_chars.is_empty() {
-            return if actual_chars.is_empty() { 0.0 } else { 1.0 };
-        }
-
-        let m = expected_chars.len();
-        let n = actual_chars.len();
-        let mut dp = vec![vec![0usize; n + 1]; m + 1];
-        for i in 0..=m {
-            dp[i][0] = i;
-        }
-        for j in 0..=n {
-            dp[0][j] = j;
-        }
-        for i in 1..=m {
-            for j in 1..=n {
-                let cost = if expected_chars[i - 1] == actual_chars[j - 1] {
-                    0
-                } else {
-                    1
-                };
-                dp[i][j] = std::cmp::min(
-                    std::cmp::min(dp[i - 1][j] + 1, dp[i][j - 1] + 1),
-                    dp[i - 1][j - 1] + cost,
-                );
-            }
-        }
-        dp[m][n] as f32 / expected_chars.len() as f32
-    }
-
-    #[test]
-    fn external_ocr_command_parser_requires_program_name() {
-        assert!(parse_external_ocr_argv("").is_err());
-        assert!(parse_external_ocr_argv("   ").is_err());
-    }
-
-    #[test]
-    fn external_ocr_command_parser_splits_args_without_shell() {
-        let argv = parse_external_ocr_argv("python3 /opt/ocr/run.py --mode fast")
-            .expect("parser should accept executable + args");
-        assert_eq!(argv[0], "python3");
-        assert_eq!(argv[1], "/opt/ocr/run.py");
-        assert_eq!(argv[2], "--mode");
-        assert_eq!(argv[3], "fast");
-    }
-
-    #[test]
-    fn overlap_sanitization_truncates_earlier_block() {
-        let mut cues = vec![
-            SubtitleCue {
-                start_ms: 1000,
-                end_ms: 3000,
-                text: "a".to_string(),
-            },
-            SubtitleCue {
-                start_ms: 2500,
-                end_ms: 4000,
-                text: "b".to_string(),
-            },
-        ];
-        sanitize_cues(&mut cues, OcrFormat::Srt);
-        assert_eq!(cues[0].end_ms, 2500);
-        assert_eq!(cues[1].start_ms, 2500);
-    }
-
-    #[test]
-    fn srt_timestamp_formats_correctly() {
-        assert_eq!(format_srt_timestamp(0), "00:00:00,000");
-        assert_eq!(format_srt_timestamp(3_723_004), "01:02:03,004");
-    }
-
-    #[test]
-    fn language_mapping_handles_iso_tags() {
-        assert_eq!(map_language_tag_to_tesseract("en"), Some("eng".to_string()));
-        assert_eq!(
-            map_language_tag_to_tesseract("eng"),
-            Some("eng".to_string())
-        );
-        assert_eq!(
-            map_language_tag_to_tesseract("pt-BR"),
-            Some("por".to_string())
-        );
-        assert_eq!(map_language_tag_to_tesseract(""), None);
-    }
-
-    #[test]
-    fn resolve_language_prefers_stream_metadata_then_config_then_system_then_english() {
-        let available = ["eng", "spa", "fra"]
-            .iter()
-            .map(|s| (*s).to_string())
-            .collect::<HashSet<_>>();
-
-        assert_eq!(
-            resolve_ocr_language(
-                Some("spa"),
-                Some("eng"),
-                Some("fr_FR.UTF-8"),
-                &available,
-                OcrEngine::Tesseract
-            ),
-            "spa"
-        );
-        assert_eq!(
-            resolve_ocr_language(
-                None,
-                Some("fra"),
-                Some("en_US.UTF-8"),
-                &available,
-                OcrEngine::Tesseract
-            ),
-            "fra"
-        );
-        assert_eq!(
-            resolve_ocr_language(
-                None,
-                None,
-                Some("fr_FR.UTF-8"),
-                &available,
-                OcrEngine::Tesseract
-            ),
-            "fra"
-        );
-        assert_eq!(
-            resolve_ocr_language(None, None, Some("zz_ZZ"), &available, OcrEngine::Tesseract),
-            "eng"
-        );
-    }
-
-    #[test]
-    fn resolve_language_ai_prefers_tags_without_lang_list() {
-        let available = HashSet::<String>::new();
-
-        assert_eq!(
-            resolve_ocr_language(Some("jpn"), None, None, &available, OcrEngine::PpOcrV4),
-            "jpn"
-        );
-        assert_eq!(
-            resolve_ocr_language(Some("eng"), None, None, &available, OcrEngine::PpOcrV3),
-            "eng"
-        );
-        assert_eq!(
-            resolve_ocr_language(None, None, None, &available, OcrEngine::External),
-            "eng"
-        );
-    }
-
-    #[test]
-    fn non_english_fallback_requires_matching_tesseract_pack() {
-        let available = ["eng"].iter().map(|s| (*s).to_string()).collect();
-        assert_eq!(
-            resolve_tesseract_fallback_language_with_available("spa", &available),
-            None
-        );
-        assert_eq!(
-            resolve_tesseract_fallback_language_with_available("fre", &available),
-            None
-        );
-    }
-
-    #[test]
-    fn english_fallback_prefers_eng_when_available() {
-        let available = ["eng"].iter().map(|s| (*s).to_string()).collect();
-        assert_eq!(
-            resolve_tesseract_fallback_language_with_available("eng", &available),
-            Some("eng".to_string())
-        );
-    }
-
-    #[test]
-    fn fallback_uses_mapped_language_code_when_available() {
-        let available = ["fra", "spa"].iter().map(|s| (*s).to_string()).collect();
-        assert_eq!(
-            resolve_tesseract_fallback_language_with_available("fre", &available),
-            Some("fra".to_string())
-        );
-        assert_eq!(
-            resolve_tesseract_fallback_language_with_available("es", &available),
-            Some("spa".to_string())
-        );
-    }
-
-    #[test]
-    fn bounding_box_to_ass_position() {
-        let bbox = OcrBoundingBox {
-            left: 80,
-            right: 120,
-            top: 790,
-            bottom: 810,
-        };
-        let (x, y) = ass_position_from_bbox(&bbox, Some((1920, 1080)));
-        assert_eq!((x, y), (100, 800));
-        let formatted = format_ass_text_with_style("Hello Bitch", Some((x, y)), None, false);
-        assert!(formatted.contains("{\\pos(100,800)}"));
-    }
-
-    #[test]
-    fn color_extraction_to_ass_hex() {
-        let color = ass_color_from_rgb(255, 0, 0);
-        assert_eq!(color, "&H0000FF&");
-        let formatted = format_ass_text_with_style("Red", None, Some((255, 0, 0)), false);
-        assert!(formatted.contains("\\c&H0000FF&"));
-    }
-
-    #[test]
-    fn dominant_color_from_rect_prefers_visible_palette() {
-        let mut pixels = vec![1u8; 4];
-        let mut palette = vec![0u8; 256 * 4];
-        palette[4] = 255; // R
-        palette[5] = 0; // G
-        palette[6] = 0; // B
-        palette[7] = 255; // A
-
-        let mut rect: ffi::AVSubtitleRect = unsafe { std::mem::zeroed() };
-        rect.w = 2;
-        rect.h = 2;
-        rect.linesize[0] = 2;
-        rect.data[0] = pixels.as_mut_ptr();
-        rect.data[1] = palette.as_mut_ptr();
-        rect.type_ = ffi::SUBTITLE_BITMAP;
-        rect.x = 0;
-        rect.y = 0;
-        rect.nb_colors = 256;
-        rect.flags = 0;
-        for i in 2..rect.data.len() {
-            rect.data[i] = ptr::null_mut();
-        }
-
-        let color = dominant_color_from_rect(&rect).expect("expected color");
-        assert_eq!(color, (255, 0, 0));
-    }
-
-    #[test]
-    fn test_model_downloads_successfully() {
-        let mut server = Server::new();
-        let body = b"dummy-onnx-model";
-        let hash = to_hex_lower(&Sha256::digest(body));
-        let mock = server
-            .mock("GET", "/model.onnx")
-            .with_status(200)
-            .with_body(body.as_slice())
-            .create();
-
-        let tmp = TempDir::new().unwrap();
-        let path = tmp.path().join("model.onnx");
-        download_model_with_values(
-            &path,
-            &format!("{}/model.onnx", server.url()),
-            &hash,
-            "model.onnx",
-        )
-        .expect("download should succeed");
-
-        mock.assert();
-        let metadata = fs::metadata(&path).unwrap();
-        assert_eq!(metadata.len(), body.len() as u64);
-    }
-
-    #[test]
-    fn test_skips_download_if_cached() {
-        let mut server = Server::new();
-        let _mock = server
-            .mock("GET", "/cached.onnx")
-            .expect(0)
-            .with_status(200)
-            .with_body("should-not-be-downloaded")
-            .create();
-
-        let tmp = TempDir::new().unwrap();
-        let path = tmp.path().join("cached.onnx");
-        let mut file = File::create(&path).unwrap();
-        file.write_all(b"cached").unwrap();
-
-        let result = ensure_model_file_with_values(
-            tmp.path(),
-            "cached.onnx",
-            &format!("{}/cached.onnx", server.url()),
-            "deadbeef",
-        );
-        assert!(result.is_ok());
-        _mock.assert();
-    }
-
-    #[test]
-    fn test_handles_corrupted_download() {
-        let mut server = Server::new();
-        let body = b"partial-data-should-fail";
-        let hash = to_hex_lower(&Sha256::digest(body));
-        let mock = server
-            .mock("GET", "/corrupt.onnx")
-            .with_status(200)
-            .with_chunked_body(|writer| {
-                writer.write_all(&body[..8])?;
-                Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "simulated drop",
-                ))
-            })
-            .create();
-
-        let tmp = TempDir::new().unwrap();
-        let path = tmp.path().join("corrupt.onnx");
-        let result = download_model_with_values(
-            &path,
-            &format!("{}/corrupt.onnx", server.url()),
-            &hash,
-            "corrupt.onnx",
-        );
-        assert!(result.is_err());
-        mock.assert();
-        assert!(!path.with_extension("download").exists());
-    }
-
-    #[test]
-    fn test_downloader_handles_404() {
-        let mut server = Server::new();
-        let mock = server
-            .mock("GET", "/missing.onnx")
-            .with_status(404)
-            .create();
-
-        let tmp = TempDir::new().unwrap();
-        let path = tmp.path().join("missing.onnx");
-        let result = download_model_with_values(
-            &path,
-            &format!("{}/missing.onnx", server.url()),
-            "deadbeef",
-            "missing.onnx",
-        );
-        assert!(result.is_err());
-        mock.assert();
-        assert!(!path.exists());
-        assert!(!path.with_extension("download").exists());
-    }
-
-    fn is_skippable_ort_runtime_error(msg: &str) -> bool {
-        let lower = msg.to_ascii_lowercase();
-        // CI runners often have an unexpected ORT runtime on PATH/LD path.
-        // Skip these environment-specific failures so OCR logic tests remain stable.
-        (lower.contains("libonnxruntime.so") && lower.contains("cannot open shared object file"))
-            || (lower.contains("onnxruntime.dll")
-                && lower.contains("is not compatible with the onnx runtime binary found"))
-            || lower.contains("expected getversionstring")
-    }
-
-    #[test]
-    fn test_onnx_session_initializes_with_fallbacks() {
-        let init_result = std::panic::catch_unwind(init_ort_environment);
-        match init_result {
-            Ok(Ok(_)) => {}
-            Ok(Err(err)) => {
-                let msg = err.to_string();
-                if is_skippable_ort_runtime_error(&msg) {
-                    eprintln!(
-                        "Skipping ORT init fallback test due to environment ORT runtime issue: {}",
-                        msg
-                    );
-                    return;
-                }
-                panic!("Failed to initialize ORT environment: {}", msg);
-            }
-            Err(payload) => {
-                let panic_msg = if let Some(msg) = payload.downcast_ref::<&str>() {
-                    *msg
-                } else if let Some(msg) = payload.downcast_ref::<String>() {
-                    msg.as_str()
-                } else {
-                    "unknown panic payload"
-                };
-                if is_skippable_ort_runtime_error(panic_msg) {
-                    eprintln!(
-                        "Skipping ORT init fallback test due to environment ORT runtime issue: {}",
-                        panic_msg
-                    );
-                    return;
-                }
-                std::panic::resume_unwind(payload);
-            }
-        }
-        assert!(
-            ORT_ENV_INIT.get().is_some(),
-            "ORT environment not initialized"
-        );
-    }
-
-    #[test]
-    fn test_gpu_requirement_env_gate() {
-        if std::env::var("DPN_OCR_REQUIRE_GPU").ok().as_deref() != Some("1") {
-            return;
-        }
-        let selection =
-            build_execution_providers().expect("GPU execution providers required but unavailable");
-        assert!(
-            selection.providers.len() > 1,
-            "Expected at least one GPU execution provider plus CPU"
-        );
-        assert!(selection.gpu_available, "GPU availability was not detected");
-    }
-
-    #[test]
-    fn test_provider_selection_prefers_cuda_when_available() {
-        let (kinds, gpu_available) =
-            select_execution_provider_plan(false, true, true, true).unwrap();
-        assert!(gpu_available);
-        assert_eq!(kinds.first(), Some(&ExecutionProviderKind::Cuda));
-        assert_eq!(kinds.last(), Some(&ExecutionProviderKind::Cpu));
-    }
-
-    #[test]
-    fn test_provider_selection_requires_gpu_flag() {
-        let err = select_execution_provider_plan(true, false, false, false)
-            .expect_err("Expected error when requiring GPU without providers");
-        assert!(
-            err.to_string().contains("No GPU execution providers"),
-            "Unexpected error: {}",
-            err
-        );
-    }
-
-    #[test]
-    fn test_auto_engine_prefers_ppocr_with_gpu() {
-        assert_eq!(
-            auto_engine_preference_with_capability(true, false),
-            OcrEngine::PpOcrV4
-        );
-    }
-
-    #[test]
-    fn test_auto_engine_prefers_ppocr_v3_on_legacy_gpu() {
-        assert_eq!(
-            auto_engine_preference_with_capability(true, true),
-            OcrEngine::PpOcrV3
-        );
-    }
-
-    #[test]
-    fn test_auto_engine_prefers_tesseract_without_gpu() {
-        assert_eq!(
-            auto_engine_preference_with_capability(false, false),
-            OcrEngine::Tesseract
-        );
-    }
-
-    #[test]
-    fn test_text_similarity_wer_like_threshold() {
-        let expected = "THIS OCR QUALITY TEST USES MANY WORDS TO ALLOW SMALL ERRORS WITHOUT FAILING STRICT THRESHOLDS IN CI RUNS TODAY ALWAYS FOR STABILITY CHECKS EACH TIME";
-        let actual = "THIS OCR QUALITY TEST USES MANY WORDS TO ALLOW SMALL ERRORS WITHOUT FAILING STRICT THRESHOLDS IN CI RUNS TODAY ALWAYS FOR STABIL1TY CHECKS EACH TIME";
-        let expected_words = normalize_text_for_word_similarity(expected);
-        let actual_words = normalize_text_for_word_similarity(actual);
-        let wer = word_error_rate(&expected_words, &actual_words);
-        let similarity = 1.0 - wer;
-        assert!(
-            similarity > 0.95,
-            "WER similarity too low: {} ({} vs {})",
-            similarity,
-            expected,
-            actual
-        );
-        let expected_chars = normalize_text_for_char_similarity(expected);
-        let actual_chars = normalize_text_for_char_similarity(actual);
-        let cer = char_error_rate(&expected_chars, &actual_chars);
-        let cer_similarity = 1.0 - cer;
-        assert!(
-            cer_similarity > 0.95,
-            "CER similarity too low: {} ({} vs {})",
-            cer_similarity,
-            expected,
-            actual
-        );
-        let jw = jaro_winkler(expected, actual);
-        assert!(jw > 0.90, "Jaro-Winkler too low: {}", jw);
-    }
-
-    #[test]
-    fn test_spatial_iou_threshold() {
-        let a = OcrBoundingBox {
-            left: 100,
-            right: 200,
-            top: 100,
-            bottom: 150,
-        };
-        let b = OcrBoundingBox {
-            left: 103,
-            right: 198,
-            top: 101,
-            bottom: 149,
-        };
-        let iou = intersection_over_union(&a, &b);
-        assert!(iou > 0.90, "IoU too low: {}", iou);
-    }
-
-    #[test]
-    fn test_ppocr_spacing_inserts_space_for_gap() {
-        let lines = vec![
-            OcrLine {
-                text: "By".to_string(),
-                bbox: Some(OcrBoundingBox {
-                    left: 0,
-                    right: 10,
-                    top: 0,
-                    bottom: 10,
-                }),
-                score: Some(0.9),
-                color: None,
-                italic: false,
-            },
-            OcrLine {
-                text: "this".to_string(),
-                bbox: Some(OcrBoundingBox {
-                    left: 14,
-                    right: 30,
-                    top: 0,
-                    bottom: 10,
-                }),
-                score: Some(0.9),
-                color: None,
-                italic: false,
-            },
-        ];
-
-        let merged = merge_ocr_lines_with_spacing(lines);
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0].text, "By this");
-    }
-
-    #[test]
-    fn test_ppocr_spacing_keeps_compact_tokens() {
-        let lines = vec![
-            OcrLine {
-                text: "By".to_string(),
-                bbox: Some(OcrBoundingBox {
-                    left: 0,
-                    right: 10,
-                    top: 0,
-                    bottom: 10,
-                }),
-                score: Some(0.9),
-                color: None,
-                italic: false,
-            },
-            OcrLine {
-                text: "this".to_string(),
-                bbox: Some(OcrBoundingBox {
-                    left: 11,
-                    right: 30,
-                    top: 0,
-                    bottom: 10,
-                }),
-                score: Some(0.9),
-                color: None,
-                italic: false,
-            },
-        ];
-
-        let merged = merge_ocr_lines_with_spacing(lines);
-        assert_eq!(merged.len(), 1);
-        assert_eq!(merged[0].text, "Bythis");
-    }
-
-    #[test]
-    fn test_ppocr_spacing_fallback_detection() {
-        let lines = vec![OcrLine {
-            text: "BythistimeIobserved".to_string(),
-            bbox: Some(OcrBoundingBox {
-                left: 0,
-                right: 100,
-                top: 0,
-                bottom: 10,
-            }),
-            score: Some(0.9),
-            color: None,
-            italic: false,
-        }];
-        assert!(ppocr_spacing_needs_fallback(&lines));
-    }
-
-    #[test]
-    fn test_quality_score_penalizes_noise() {
-        let clean = "By this time, I observed that the rain had stopped.";
-        let noisy = "Bythistime,I0bserved @&| the rain had st0pped";
-        let clean_score = ocr_text_quality_score(clean, "eng");
-        let noisy_score = ocr_text_quality_score(noisy, "eng");
-        assert!(
-            clean_score > noisy_score,
-            "expected clean score ({clean_score}) > noisy score ({noisy_score})"
-        );
-    }
-
-    #[test]
-    fn test_quality_fallback_detection_triggers_on_noisy_text() {
-        let lines = vec![OcrLine {
-            text: "BythistimeI0bserved@&|".to_string(),
-            bbox: Some(OcrBoundingBox {
-                left: 0,
-                right: 100,
-                top: 0,
-                bottom: 10,
-            }),
-            score: Some(0.95),
-            color: None,
-            italic: false,
-        }];
-        assert!(ppocr_needs_quality_fallback(&lines, "eng"));
-    }
-
-    #[test]
-    fn test_quality_fallback_detection_avoids_good_english_text() {
-        let lines = vec![OcrLine {
-            text: "By this time, I observed the village from afar.".to_string(),
-            bbox: Some(OcrBoundingBox {
-                left: 0,
-                right: 200,
-                top: 0,
-                bottom: 12,
-            }),
-            score: Some(0.93),
-            color: None,
-            italic: false,
-        }];
-        assert!(!ppocr_needs_quality_fallback(&lines, "eng"));
-    }
-
-    #[test]
-    fn test_postprocess_english_glue_and_punctuation() {
-        let src = "ConstableCrane? Notonlyme. beforehewentintotheriver";
-        let got = postprocess_ocr_text(src, "eng");
-        assert_eq!(
-            got,
-            "Constable Crane? Not only me. before he went into the river"
-        );
-    }
-
-    #[test]
-    fn test_postprocess_english_deglues_common_tokens() {
-        let src = "Ibegpardon. Standdown! Ihavenot Loppedoff? Ishall return in the l9th century.";
-        let got = postprocess_ocr_text(src, "eng");
-        assert_eq!(
-            got,
-            "I beg pardon. Stand down! I have not Lopped off? I shall return in the 19th century."
-        );
-    }
-
-    #[test]
-    fn test_postprocess_english_fixes_silence_glue_cases() {
-        let src = "to be as dark as itis now. Whylost? He'salive?";
-        let got = postprocess_ocr_text(src, "eng");
-        assert_eq!(got, "to be as dark as it is now. Why lost? He's alive?");
-    }
-
-    #[test]
-    fn test_split_glued_token_dp_segmentation() {
-        assert_eq!(
-            split_glued_ascii_token("Ibegpardon").as_deref(),
-            Some("I beg pardon")
-        );
-        assert_eq!(
-            split_glued_ascii_token("Standdown").as_deref(),
-            Some("Stand down")
-        );
-        assert_eq!(split_glued_ascii_token("Tonight"), None);
-    }
-
-    #[test]
-    fn test_postprocess_non_english_passthrough() {
-        let src = "Notonlyme";
-        let got = postprocess_ocr_text(src, "jpn");
-        assert_eq!(got, "Notonlyme");
-    }
-
-    #[test]
-    fn test_color_distance_threshold() {
-        let mut pixels = vec![1u8; 4];
-        let mut palette = vec![0u8; 256 * 4];
-        palette[4] = 255;
-        palette[5] = 0;
-        palette[6] = 0;
-        palette[7] = 255;
-
-        let mut rect: ffi::AVSubtitleRect = unsafe { std::mem::zeroed() };
-        rect.w = 2;
-        rect.h = 2;
-        rect.linesize[0] = 2;
-        rect.data[0] = pixels.as_mut_ptr();
-        rect.data[1] = palette.as_mut_ptr();
-        rect.type_ = ffi::SUBTITLE_BITMAP;
-        rect.x = 0;
-        rect.y = 0;
-        rect.nb_colors = 256;
-        rect.flags = 0;
-        for i in 2..rect.data.len() {
-            rect.data[i] = ptr::null_mut();
-        }
-
-        let color = dominant_color_from_rect(&rect).expect("expected color");
-        let distance = rgb_distance(color, (255, 0, 0));
-        assert!(distance < 15.0, "distance too high: {}", distance);
-    }
-
-    #[test]
-    fn test_golden_dataset_quality() {
-        if std::env::var("DPN_OCR_GOLDEN").ok().as_deref() != Some("1") {
-            eprintln!("Skipping golden OCR quality test (set DPN_OCR_GOLDEN=1 to enable).");
-            return;
-        }
-
-        let dataset_dir = PathBuf::from("tests/golden_subs");
-        if !dataset_dir.is_dir() {
-            eprintln!("Golden dataset directory not found; skipping.");
-            return;
-        }
-
-        let model_dir = match resolve_model_dir() {
-            Ok(dir) => dir,
-            Err(err) => {
-                eprintln!("Model dir unavailable: {err}. Skipping.");
-                return;
-            }
-        };
-
-        let engine_result =
-            std::panic::catch_unwind(|| PpOcrEngine::new(&model_dir, PpOcrVariant::V4, false));
-        let mut engine = match engine_result {
-            Ok(Ok(engine)) => engine,
-            Ok(Err(err)) => {
-                eprintln!("OCR engine unavailable: {err}. Skipping.");
-                return;
-            }
-            Err(payload) => {
-                let panic_msg = if let Some(msg) = payload.downcast_ref::<&str>() {
-                    *msg
-                } else if let Some(msg) = payload.downcast_ref::<String>() {
-                    msg.as_str()
-                } else {
-                    "unknown panic payload"
-                };
-                if is_skippable_ort_runtime_error(panic_msg) {
-                    eprintln!(
-                        "Skipping golden OCR quality test due to environment ORT runtime issue: {}",
-                        panic_msg
-                    );
-                    return;
-                }
-                std::panic::resume_unwind(payload);
-            }
-        };
-
-        #[derive(serde::Deserialize)]
-        struct GoldenExpected {
-            expected_text: String,
-            expected_bbox: [i32; 4],
-            expected_color_rgb: [u8; 3],
-            is_italic: Option<bool>,
-        }
-
-        let mut found = 0usize;
-        for entry in fs::read_dir(&dataset_dir).unwrap() {
-            let entry = entry.unwrap();
-            let path = entry.path();
-            if path.extension().and_then(|e| e.to_str()) != Some("png") {
-                continue;
-            }
-            let json_path = path.with_extension("json");
-            if !json_path.exists() {
-                continue;
-            }
-            found += 1;
-            let json = fs::read_to_string(&json_path).unwrap();
-            let expected: GoldenExpected = serde_json::from_str(&json).unwrap();
-
-            let output = engine.extract_lines(&path, "eng").unwrap();
-            let Some(line) = output.lines.first() else {
-                panic!("No OCR output for {:?}", path);
-            };
-
-            let expected_text = expected.expected_text.trim();
-            let actual_text = line.text.trim();
-            let expected_words = normalize_text_for_word_similarity(expected_text);
-            let actual_words = normalize_text_for_word_similarity(actual_text);
-            let wer = word_error_rate(&expected_words, &actual_words);
-            let expected_chars = normalize_text_for_char_similarity(expected_text);
-            let actual_chars = normalize_text_for_char_similarity(actual_text);
-            let cer = char_error_rate(&expected_chars, &actual_chars);
-            let similarity = (1.0 - wer).max(1.0 - cer);
-            assert!(
-                similarity > 0.95,
-                "Text similarity too low for {:?}: {} (wer {:.3}, cer {:.3})",
-                path,
-                similarity,
-                wer,
-                cer
-            );
-
-            let expected_bbox = OcrBoundingBox {
-                left: expected.expected_bbox[0],
-                top: expected.expected_bbox[1],
-                right: expected.expected_bbox[2],
-                bottom: expected.expected_bbox[3],
-            };
-            let Some(actual_bbox) = line.bbox.as_ref() else {
-                panic!("Missing bounding box for {:?}", path);
-            };
-            let iou = intersection_over_union(&expected_bbox, actual_bbox);
-            assert!(iou > 0.90, "IoU too low for {:?}: {}", path, iou);
-
-            let expected_color = (
-                expected.expected_color_rgb[0],
-                expected.expected_color_rgb[1],
-                expected.expected_color_rgb[2],
-            );
-            if let Some(actual_color) = line.color {
-                let dist = rgb_distance(actual_color, expected_color);
-                assert!(
-                    dist < 15.0,
-                    "Color distance too high for {:?}: {}",
-                    path,
-                    dist
-                );
-            }
-
-            if let Some(italic) = expected.is_italic {
-                assert_eq!(line.italic, italic, "Italic mismatch for {:?}", path);
-            }
-        }
-
-        if found == 0 {
-            eprintln!("Golden dataset empty; skipping.");
-        }
-    }
-
-    #[test]
-    #[ignore]
-    fn test_manual_ppocr_v3_single_image_probe() {
-        if std::env::var("DPN_OCR_MANUAL_PPOCR_V3").ok().as_deref() != Some("1") {
-            eprintln!("Skipping manual PP-OCRv3 probe (set DPN_OCR_MANUAL_PPOCR_V3=1 to enable).");
-            return;
-        }
-
-        let model_dir = resolve_model_dir().expect("resolve model dir");
-        let gpu_available = init_ort_environment().expect("init ORT environment");
-        eprintln!(
-            "ORT environment initialized; gpu_available={}",
-            gpu_available
-        );
-
-        let mut engine = init_ppocr_engine(&model_dir, require_gpu(), PpOcrVariant::V3)
-            .expect("init PP-OCRv3 engine");
-
-        let tmp_dir = tempfile::tempdir().expect("create temp dir");
-        let image_path = tmp_dir.path().join("manual_probe.png");
-        let img = image::RgbImage::from_pixel(1280, 720, image::Rgb([0, 0, 0]));
-        img.save(&image_path).expect("save probe image");
-
-        let output = engine
-            .extract_lines(&image_path, "eng")
-            .expect("run PP-OCRv3 inference");
-        let (sum_conf, count_conf) = output
-            .lines
-            .iter()
-            .filter_map(|line| line.score)
-            .fold((0.0f32, 0usize), |(sum, count), score| {
-                (sum + score, count + 1)
-            });
-        let avg_conf = if count_conf > 0 {
-            sum_conf / count_conf as f32
-        } else {
-            0.0
-        };
-        eprintln!(
-            "PP-OCRv3 probe completed: lines={}, avg_conf={:.4}",
-            output.lines.len(),
-            avg_conf
-        );
-    }
-}
+mod tests;
